+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -301,8 +301,6 @@ func (backend *MetricsReportingKurtosisBackend) CreateUserService(
 	entrypointArgs []string,
 	cmdArgs []string,
 	envVars map[string]string,
-	enclaveDataDirpathOnHostMachine string,
-	enclaveDataDirpathOnContainer string,
 	filesArtifactMountDirpaths map[string]string,
 ) (
 	newUserService *service.Service,
@@ -319,25 +317,23 @@ func (backend *MetricsReportingKurtosisBackend) CreateUserService(
 		entrypointArgs,
 		cmdArgs,
 		envVars,
-		enclaveDataDirpathOnHostMachine,
-		enclaveDataDirpathOnContainer,
 		filesArtifactMountDirpaths,
 	)
 	if err != nil {
-		return nil,
-			stacktrace.Propagate(
-				err,
-				"An error occurred creating the user service with ID '%v' and GUID '%v' using image '%v' with private ports '%+v' with entry point args '%+v', command args '%+v', environment vars '%+v', enclave data mount dirpath '%v' and file artifacts mount dirpath '%v'",
-				id,
-				guid,
-				containerImageName,
-				privatePorts,
-				entrypointArgs,
-				cmdArgs,
-				envVars,
-				enclaveDataDirpathOnHostMachine,
-				filesArtifactMountDirpaths,
-			)
+		return nil, stacktrace.Propagate(
+			err,
+			"An error occurred creating the user service with ID '%v' and GUID '%v' using image '%v' " +
+				"with private ports '%+v' with entry point args '%+v', command args '%+v', environment " +
+				"vars '%+v', and file artifacts mount dirpath '%v'",
+			id,
+			guid,
+			containerImageName,
+			privatePorts,
+			entrypointArgs,
+			cmdArgs,
+			envVars,
+			filesArtifactMountDirpaths,
+		)
 	}
 	return userService, nil
 }