+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -35,8 +35,15 @@ func (backend *MetricsReportingKurtosisBackend) PullImage(image string) error {
 	return nil
 }
 
-func (backend *MetricsReportingKurtosisBackend) CreateEngine(ctx context.Context, imageOrgAndRepo string, imageVersionTag string, grpcPortNum uint16, grpcProxyPortNum uint16, engineDataDirpathOnHostMachine string, envVars map[string]string) (*engine.Engine, error) {
-	result, err := backend.underlying.CreateEngine(ctx, imageOrgAndRepo, imageVersionTag, grpcPortNum, grpcProxyPortNum, engineDataDirpathOnHostMachine, envVars)
+func (backend *MetricsReportingKurtosisBackend) CreateEngine(ctx context.Context, imageOrgAndRepo string, imageVersionTag string, grpcPortNum uint16, grpcProxyPortNum uint16, envVars map[string]string) (*engine.Engine, error) {
+	result, err := backend.underlying.CreateEngine(
+		ctx,
+		imageOrgAndRepo,
+		imageVersionTag,
+		grpcPortNum,
+		grpcProxyPortNum,
+		envVars,
+	)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred creating the engine using image '%v' with tag '%v'", imageOrgAndRepo, imageVersionTag)
 	}
@@ -206,7 +213,6 @@ func (backend *MetricsReportingKurtosisBackend) CreateModule(
 	guid module.ModuleGUID,
 	ipAddr net.IP, // TODO REMOVE THIS ONCE WE FIX THE STATIC IP PROBLEM!!
 	grpcPortNum uint16,
-	enclaveDataDirpathOnHostMachine string,
 	envVars map[string]string,
 ) (
 	newModule *module.Module,
@@ -220,7 +226,6 @@ func (backend *MetricsReportingKurtosisBackend) CreateModule(
 		guid,
 		ipAddr,
 		grpcPortNum,
-		enclaveDataDirpathOnHostMachine,
 		envVars,
 	)
 	if err != nil {
@@ -301,8 +306,6 @@ func (backend *MetricsReportingKurtosisBackend) CreateUserService(
 	entrypointArgs []string,
 	cmdArgs []string,
 	envVars map[string]string,
-	enclaveDataDirpathOnHostMachine string,
-	enclaveDataDirpathOnContainer string,
 	filesArtifactMountDirpaths map[string]string,
 ) (
 	newUserService *service.Service,
@@ -319,25 +322,23 @@ func (backend *MetricsReportingKurtosisBackend) CreateUserService(
 		entrypointArgs,
 		cmdArgs,
 		envVars,
-		enclaveDataDirpathOnHostMachine,
-		enclaveDataDirpathOnContainer,
 		filesArtifactMountDirpaths,
 	)
 	if err != nil {
-		return nil,
-			stacktrace.Propagate(
-				err,
-				"An error occurred creating the user service with ID '%v' and GUID '%v' using image '%v' with private ports '%+v' with entry point args '%+v', command args '%+v', environment vars '%+v', enclave data mount dirpath '%v' and file artifacts mount dirpath '%v'",
-				id,
-				guid,
-				containerImageName,
-				privatePorts,
-				entrypointArgs,
-				cmdArgs,
-				envVars,
-				enclaveDataDirpathOnHostMachine,
-				filesArtifactMountDirpaths,
-			)
+		return nil, stacktrace.Propagate(
+			err,
+			"An error occurred creating the user service with ID '%v' and GUID '%v' using image '%v' " +
+				"with private ports '%+v' with entry point args '%+v', command args '%+v', environment " +
+				"vars '%+v', and file artifacts mount dirpath '%v'",
+			id,
+			guid,
+			containerImageName,
+			privatePorts,
+			entrypointArgs,
+			cmdArgs,
+			envVars,
+			filesArtifactMountDirpaths,
+		)
 	}
 	return userService, nil
 }