+++ container-engine-lib/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Breaking Changes
+* Removed `enclaveDataDirpathOnHostMachine` and `enclaveDataDirpathOnServiceContainer` from `KurtosisBackend.CreateUserService`
+    * Users no longer need to provide this argument
 
 # 0.18.0
 ### Features