+++ container-engine-lib/docs/changelog.md
@@ -1,4 +1,11 @@
 # TBD
+### Breaking Changes
+* Removed `enclaveDataDirpathOnHostMachine` and `enclaveDataDirpathOnServiceContainer` from `KurtosisBackend.CreateUserService`
+    * Users no longer need to provide this argument
+* Removed `enclaveDataDirpathOnHostMachine` argument from `KurtosisBackend.CreateModule`
+    * Users no longer need to provide this argument
+* Removed `engineDataDirpathOnHostMachine` from `KurtosisBackend.CreateEngine`
+    * Users no longer need to provide this argument
 
 # 0.18.0
 ### Features